其次manager每次通过一个compensation leave以后 需要响应减去4小时的用户加班时间 每次撤销一个已经通过的compensation leave需要把时间重新加上去
判断逻辑，如果通过了一个medical leave 需要加上该employee的medical leave时间

判断逻辑，如果通过了一个annual leave 需要禁用当年的entitlement to annual leave

如果cancel compensation leave的话会重新恢复employee的overworking time

如果是cancel medical leave的话需要重新减去它的总medical时间

如果是cancel annual leave会重新恢复entitlement

(done 已解决 🍾)

默认禁止老板发起申请，其他每个发起申请的人都有上级 （done）

返回所有需要user_id审批的申请，既然需要他审批，说明受到他的直接管理，而不是越级。

(done 已解决 🍾)

员工发起annual leave申请的时候，按照员工类型，获取最大的年假时间，不能超过这个点

(done 已解决 🍾)

+++ CA/lapsBackend/src/main/java/com/batch57/gdipsa/group6/lapsbackend/controller/application/ApplicationController.java
@@ -8,6 +8,7 @@ import com.batch57.gdipsa.group6.lapsbackend.model.enumLayer.OVERWORKING_UNIT;
 import com.batch57.gdipsa.group6.lapsbackend.model.user.employee.model.Employee;
 import com.batch57.gdipsa.group6.lapsbackend.serviceLayer.application.ApplicationInterfaceImplementation;
 import com.batch57.gdipsa.group6.lapsbackend.serviceLayer.application.CompensationLeaveValidator;
+import com.batch57.gdipsa.group6.lapsbackend.serviceLayer.department.DepartmentInterfaceImplementation;
 import com.batch57.gdipsa.group6.lapsbackend.serviceLayer.user.employeeInterfaceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +19,10 @@ import org.springframework.validation.Validator;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import static org.aspectj.runtime.internal.Conversions.booleanValue;
@@ -28,8 +32,8 @@ import static org.aspectj.runtime.internal.Conversions.booleanValue;
  * 通常被Java的LocalDateTime所接受。ISO 8601的日期时间格式类似于这样："2023-12-31T23:59:59"
  *
  * {
- *     "fromDate": "2023-12-20T08:00:00",
- *     "toDate": "2023-12-25T17:00:00",
+ *     "fromDate": "2023-12-20",
+ *     "toDate": "2023-12-25",
  *     "employeeLeaveType": "ANNUAL_LEAVE"
  * }
  *
@@ -46,7 +50,8 @@ public class ApplicationController {
     private employeeInterfaceImpl employeeService;
     @Autowired
     private ApplicationInterfaceImplementation applicationService;
-
+    @Autowired
+    private DepartmentInterfaceImplementation departmentService;
 
     // compensation检查器
     /**
@@ -83,6 +88,10 @@ public class ApplicationController {
         Employee employee = employeeService.GetEmployeeById(user_id);
         if(employee == null) return new ResponseEntity<>("ERROR in binding employee", HttpStatus.NOT_FOUND);
 
+        // 没有上级的人不允许发起申请
+        Employee superior = employeeService.GetSuperior(user_id);
+        if(superior == null) return new ResponseEntity<>("You dont't have a superior, submitting an application is not allowed", HttpStatus.EXPECTATION_FAILED);
+
         // 申请信息有错误
         if (bindingResult.hasErrors()) {
             bindingResult.getAllErrors().forEach(error -> {
@@ -98,13 +107,25 @@ public class ApplicationController {
 
         // 检查该用户有无申请资格
         boolean isApplicable = booleanValue(CheckIfEmployeeIsApplicable(user_id).getBody()); // 检查有无已经提交的申请
+        if(!isApplicable){
+            return new ResponseEntity<>("You have alive application waited to be viewed, please contact your direct manager", HttpStatus.EXPECTATION_FAILED);
+        }
+
+        // 检查用户是否有资格申请annual leave
         isApplicable = newApplication.getEmployeeLeaveType() == EMPLOYEE_LEAVE_TYPE.ANNUAL_LEAVE? isApplicable && employee.isEntitlementToAnnualLeave() : isApplicable; // 检查是否entitled to annul leave
         if(!isApplicable) {
-            return new ResponseEntity<>("Your are not applicable to this type of leave.", HttpStatus.EXPECTATION_FAILED);
+            return new ResponseEntity<>("Your are not applicable to annual leave because you have already applied for it.", HttpStatus.EXPECTATION_FAILED);
+        }
+        // 如果有资格申请annual leave的话，检查是否超过最大允许年假时间
+        if(newApplication.getEmployeeLeaveType() == EMPLOYEE_LEAVE_TYPE.ANNUAL_LEAVE) {
+            int maximum_annual_leave = employee.getEmployeeType().getAnnualLeave();
+            if(maximum_annual_leave < newApplication.getDayOff()) {
+                return new ResponseEntity<>("Your maximum annual leave is " + maximum_annual_leave + " days", HttpStatus.EXPECTATION_FAILED);
+            }
         }
 
         // 如果是compensation 的话，检查有无超过最大的允许unit：overworking/unit_time
-        if(isExceedMaximumCompensationUnit(employee, newApplication)){
+        if(newApplication.getEmployeeLeaveType()==EMPLOYEE_LEAVE_TYPE.COMPENSATION_LEAVE && isExceedMaximumCompensationUnit(employee, newApplication)){
             return new ResponseEntity<>("Day off exceeds maximum compensation units allowed.", HttpStatus.EXPECTATION_FAILED);
         }
 
@@ -125,7 +146,6 @@ public class ApplicationController {
         }
     }
 
-
     /**
      * 获取特定的application信息
      *
@@ -144,19 +164,18 @@ public class ApplicationController {
         }
     }
 
-
     /**
      * 根据编号删除application，如果成功，返回被删除对象。如果已经approved了，不能删除
      * @param application_id
      * @return
      */
     @DeleteMapping("/delete/{application_id}")
-    public ResponseEntity<Application> DeleteApplicationById(@PathVariable("application_id") int application_id) {
+    public ResponseEntity<?> DeleteApplicationById(@PathVariable("application_id") int application_id) {
         Application application = applicationService.GetApplicationById(application_id);
 
         // 通过了就不能删除
-        if(application.getApplicationStatus() == APPLICATION_STATUS.APPROVED || application.getApplicationStatus() == APPLICATION_STATUS.CANCELLED) {
-            return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+        if(application.getApplicationStatus() == APPLICATION_STATUS.APPROVED || application.getApplicationStatus() == APPLICATION_STATUS.CANCELLED || application.getApplicationStatus() == APPLICATION_STATUS.REJECTED) {
+            return new ResponseEntity<>("Current status " + application.getApplicationStatus() + " can't be deleted",  HttpStatus.EXPECTATION_FAILED);
         }
 
         Application updated = applicationService.DeleteApplicationById(application_id);
@@ -167,7 +186,6 @@ public class ApplicationController {
             return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
         }
     }
-
 
     /**
      * 返回某个用户id的所有假期申请
@@ -184,6 +202,12 @@ public class ApplicationController {
         }
     }
 
+    /**
+     * 返回某个用户某个状态的所有申请
+     * @param user_id
+     * @param status
+     * @return
+     */
     @GetMapping("/get-application-by-employee-id-status/{user_id}/{status}")
     public ResponseEntity<List<Application>> GetApplicationByEmployeeIdAndStatus(@PathVariable("user_id") int user_id, @PathVariable("status") APPLICATION_STATUS status) {
         List<Application> applications = GetApplicationByEmployeeId(user_id).getBody();
@@ -211,7 +235,6 @@ public class ApplicationController {
         }
     }
 
-
     /**
      * 会把status自动解析成APPLICATION_STATUS，
      * 如果失败直接返回失败
@@ -233,7 +256,7 @@ public class ApplicationController {
 
     /**
      * 检查某个用户是否有资格提交申请
-     * 判断条件是有没有活跃的申请
+     * 判断条件是有没有活跃的申请[appiled/updated]
      * @param user_id
      * @return
      */
@@ -247,6 +270,10 @@ public class ApplicationController {
         return ResponseEntity.ok(!isApplicable);
     }
 
+    @GetMapping("/get-subordinates-alive-application")
+    public ResponseEntity<?> GetSubordinatesAliveApplication(@RequestHeader("user_id") int user_id) {
+        return new ResponseEntity<>(GetAliveApplicationLedBy(user_id), HttpStatus.OK);
+    }
 
     /**
      * 以下逻辑处理一个employee最多能申请几个compensation leave unit
@@ -264,4 +291,79 @@ public class ApplicationController {
 
         return false;
     }
+
+    /**
+     * 返回某个人领导下的所有申请
+     */
+    public List<Application> GetAliveApplicationLedBy(int user_id){
+        Employee employee = employeeService.GetEmployeeById(user_id);
+
+        // 判断领导哪个部门
+        Integer leads_department = employeeService.isManager(user_id);
+        if(leads_department == -1) {
+            return null;
+        }
+
+        // 获取手下所有员工 + 子部门的领导
+        List<Employee> employees = departmentService.GetEmployeesAndSubManagerByDepartmentId(leads_department);
+        Set<Integer> idSet = new HashSet<>();
+        employees
+                .forEach(e -> {
+                    idSet.add(e.getUser_id());
+                });
+
+        // 筛选这些员工的alive application
+        List<Application> allApplications = applicationService.GetAllApplication();
+        allApplications= allApplications
+                .stream()
+                .filter( a -> idSet.contains(a.getEmployee().getUser_id()) && a.getEmployee().getUser_id()!=user_id)
+                .toList();
+
+        // 筛选里面的活跃请求
+        allApplications = allApplications
+                .stream()
+                .filter(a -> a.getApplicationStatus()==APPLICATION_STATUS.APPLIED || a.getApplicationStatus()==APPLICATION_STATUS.UPDATED)
+                .toList();
+
+        return allApplications;
+    }
+
+    /**
+     * 返回的申请里不包括cancelled 和 approved 和 rejected
+     * @param user_id
+     * @return
+     */
+    @GetMapping("/get-application-waited-to-be-viewed-by")
+    public ResponseEntity<?> GetApplicationWaitedToBeViewedByUserId(@RequestHeader("user_id") int user_id) {
+        List<Application> applications = applicationService.GetApplicationWaitedToBeViewedByUserId(user_id);
+        applications = applications
+                .stream()
+                .filter(a -> a.getApplicationStatus()==APPLICATION_STATUS.APPLIED || a.getApplicationStatus()==APPLICATION_STATUS.UPDATED)
+                .toList();
+        if(applications == null) {
+            return new ResponseEntity<>("There is no alive application waited to be viewed", HttpStatus.NOT_FOUND);
+        }else {
+            return new ResponseEntity<>(applications, HttpStatus.OK);
+        }
+    }
+
+    /**
+     * 返回的申请里不包括cancelled 和 approved 和 rejected
+     * @param user_id
+     * @return
+     */
+    @GetMapping("/get-approved-application-to-be-viewed-by")
+    public ResponseEntity<?> GetApprovedApplicationToBeViewedByUserId(@RequestHeader("user_id") int user_id) {
+        List<Application> applications = applicationService.GetApplicationWaitedToBeViewedByUserId(user_id);
+        applications = applications
+                .stream()
+                .filter(a -> a.getApplicationStatus()==APPLICATION_STATUS.APPROVED)
+                .toList();
+        if(applications == null) {
+            return new ResponseEntity<>("There is no alive application waited to be viewed", HttpStatus.NOT_FOUND);
+        }else {
+            return new ResponseEntity<>(applications, HttpStatus.OK);
+        }
+    }
+
 }
